Student Grade Calculator
It is a program that prompts the user to input student marks. The input should be between 0 and 100. Then output the correct grade:

A > 79
B - 60 to 79
C -  49 to 59
D - 40 to 49
E - less 40.

You use git clone to be able to download the documents in the GitHub

Git

Git clone git@github.com:OkemwaEnid/First-code-challenge.git